Consolidate modal components and update imports
Moved modal components into src/components/modal, updated related import paths, and fixed TypeScript check issue.
